1st_commit
This is project made in Mathematica to create a fast and easy tool to
show the actual air pollution among the world. The data are taken from
the website aqicn.org/api/ which provides air pollution information from
9000 air measurement stations around the world. The geolocalization data
is taken from google api. Please feel free to distribute it and to
develop. You can also use the wolfram cloud deployment by this link:
https://www.wolframcloud.com/objects/5d02a3ff-6611-4d16-82b0-e9f4cb348d3b